Split out getPostEdits
